refactor(ipod): extract useIpodStatusBacklight from useIpodLogic
